requirements.txt
flask
langchain
langchain-openai
python-dotenv
pypdf
